// src/generated/models/Pulse_modulesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pulse_modulesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pulse_modulesstatecode = keyof typeof Pulse_modulesstatecode;
export const Pulse_modulesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pulse_modulesstatuscode = keyof typeof Pulse_modulesstatuscode;

export interface Pulse_modulesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pulse_colorcode?: string;
  pulse_description?: string;
  pulse_iconurl?: string;
  pulse_isactive?: boolean;
  pulse_moduleid: string;
  pulse_name?: string;
  pulse_order?: number;
  statecode: Pulse_modulesstatecode;
  statuscode?: Pulse_modulesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pulse_modules extends Pulse_modulesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pulse_isactivename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}